requirements.txt
numpy==1.16.5
onnxruntime==1.0.0
scipy==1.3.1
tqdm==4.36.1
Pillow==6.2.1
scikit_learn==0.22
